# Dyno: improve performance when resolving