--- a/frontend/lib/resolution/InitResolver.cpp
+++ b/frontend/lib/resolution/InitResolver.cpp
@@ -143,7 +143,7 @@ void InitResolver::resolveImplicitSuperInit() {
     // Capture the errors emitted here and defer them until we know we haven't
     // found a "real" super.init call (which means a different error supercedes
     // these).
-    auto cAndErrors = ctx_->runAndTrackErrors([&c](Context* ctx) {
+    auto cAndErrors = ctx_->runAndCaptureErrors([&c](Context* ctx) {
       c.noteResultPrintCandidates(nullptr);
       return true;
     });