… frontend/lib/parsing/parsing-queries.cpp
… frontend/lib/resolution/Resolver.cpp
… frontend/lib/resolution/prims.cpp
… frontend/lib/resolution/resolution-queries.cpp
… frontend/test/framework/testErrorTracking.cpp
… frontend/test/resolution/testCompilerError.cpp