This PR addresses an issue I observed while working on https://github.com/chapel-lang/chapel/pull/27031, namely that resolving `blockDist` was _slow_. After doing some profiling, I observed that a big chunk of time is spent in `storeErrorsFor`, which is a function invoked for `runAndTrackErrors`. My first course of actions was to disable unnecessary calls to `runAndTrackErrors`, which occurred in https://github.com/chapel-lang/chapel/pull/27041. This helped (see below table), but I still observed calls to `storeErrors` when profiling.

This call is unfortunately necessary for correctness: collecting errors needs to work in the presence of query caching, which means that we need to be able to capture errors even if we aren't actually re-running a computation. This is done by traversing the dependency graph to collect such errors. The traversal incurs a performance penalty, despite existing optimizations (skipping sub-graphs that don't have errors). 

The optimization in this PR comes from the observation that although collecting errors incurs this performance penalty, most callers to `runAndTrackErrors` do not care about the list of errors; they only care about whether these errors have occurred. Given that we already track whether errors occurred in sub-graphs (the optimization mentioned in the previous paragraph), this means that for calls that only care about the presence of errors, we can produce a result in O(1) time. This PR does just that.

A minor change to the optimization was to distinguish errors from warnings. `ranWithoutErrors` returns `ranWithoutErrors=true` if only errors were encountered; however, the Context tracks whether _any_ `Error`s were emitted, which includes warnings. Thus, this PR turns the boolean into a two-bit field, storing the presences of warnings and errors separately. 

While there, I noticed that `errorsPresentInSelfOrDependencies` was updated too aggressively, and did not account for silencing. As a result, some subgraphs were marked as having errors despite not having any. This previously simply reduced the number of cases in which the graph-skipping optimization fired, but now is important for correctness. This PR fixes that.

Finally, I noticed that there were some overheads in the profiler output from using a regular `set` from the standard library for tracking recursion errors. Recursion errors are the exception rather than the norm; moreover, the set being that of pointers enables some optimizations. I switched the code to use `llvm::SmallPtrSet`, which had a modest performance boost of 3-6%. 

## Performance Impact
| __Experiment__                                                                   | Time (debug) | Time (release) |
|----------------------------------------------------------------------------------|--------------|----------------|
| Baseline                                                                         | 91 seconds   | 9.7 seconds    |
| Remove runAndTrackErrors (#27041)                                                | 42 seconds   | -              |
| Stop collecting errors where possible (4db46d4c5ea463a5ea7c648cd5387e9a2a98e89f) | 31 seconds   | -              |
| Switch to llvm::SmallPtrSet (2e8f3477b1d6eccf7ed4c76145751105b395848d)           | 29 seconds   | 4.0 seconds    |

Reviewed by @arifthpe -- thanks!

## Testing
- [x] dyno tests
- [x] paratest